+++ CHANGELOG.md
@@ -5,7 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 ### Fixed
 - fix salutation in order an customer
-- paypal invoice and paypal instalment payment data was discarded, the data is now transfered correctly
+- paypal invoice and paypal installment payment data was discarded, the data is now transfered correctly
 
 ### Changed
 - corrected the name of the isMappedIdentity function of the IdentityService